Update /ghauth message to include `oauth_code:`
If a user copies the full command with token, Discord won't recognise it
as a slash command, causing the message to get posted to the channel.
This is not good. Turns out that to have it recognised with the token,
the arg name needs to be provided. Adding `oauth_code:` to the command
to copy results in Discord correctly identifying it as a slash command,
and it's filled out correctly.