Added urls for searching
I'm definitely not happy with the current state of it as the goal is to fix https://github.com/runelite/runelite.net/issues/461 and current state of things don't use query parameters.

I've tried getting query parameters to work but for some reason I can't seem to figure out how to match the url for those.

So I went ahead and started trying to 'solve the issue' this way.

I would like to get some feedback and/or help on how to tackle this problem.